Refactor: 뉴스 리스트 컴포넌트 리팩토링
