+++ src/widgets/post/components/NewsListHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './NewsListHeader';
