+++ src/widgets/post/sections/NewsList/index.ts
@@ -1,0 +1,1 @@
+export * from './NewsList';
